Add unit tests and based on test modifies actual code
    Unit test saved potential future bug from arisen, since (int)
    type coercing forced 0X values to become only X, which obviously
    causes errors inside DateTime methods.

+++ components/MyComponent.php
@@ -7,7 +7,7 @@ class MyComponent extends \yii\base\Component
     public static function getAgeFromPersonCode($personalCode) {
         $dobIdentifier = substr($personalCode, 1, 6);
 
-        $year = (int) substr($dobIdentifier, 0, 2);
+        $year = substr($dobIdentifier, 0, 2);
         $year = ($year > 16) ? 19 . $year : 20 . $year;
 
         $month = substr($dobIdentifier, 2, 2);
